requirements.txt
baidu-aip
python-telegram-bot
pymysql
